+++ dashboard-server/src/test/java/com/footballstatsdashboard/resources/ClubResourceTest.java
@@ -297,49 +297,6 @@ public class ClubResourceTest {
     }
 
     /**
-     * given that the request contains a club that does not belong to the current user, tests that the club data is not
-     * updated and a service exception is thrown instead
-     */
-    @Test
-    public void updateClubWhenClubDoesNotBelongToUser() {
-        // setup
-        UUID existingClubId = UUID.randomUUID();
-        Club existingClub = ClubDataProvider.ClubBuilder.builder()
-                .isExisting(true)
-                .existingUserId(userPrincipal.getId())
-                .withId(existingClubId)
-                .withIncome()
-                .withExpenditure()
-                .build();
-        when(clubService.getClub(eq(existingClubId), eq(userPrincipal.getId())))
-                .thenThrow(new ServiceException(HttpStatus.NOT_FOUND_404, "No club found!"));
-
-        BigDecimal updatedWageBudget = existingClub.getWageBudget().add(new BigDecimal("100"));
-        BigDecimal updatedTransferBudget = existingClub.getTransferBudget().add(new BigDecimal("100"));
-        BigDecimal totalFunds = updatedTransferBudget.add(updatedWageBudget);
-        Club incomingClubBase = ClubDataProvider.ClubBuilder.builder()
-                .isExisting(false)
-                .withId(existingClubId)
-                .build();
-        Club incomingClub = ClubDataProvider.ModifiedClubBuilder.builder()
-                .from(incomingClubBase)
-                .withUpdatedTransferBudget(updatedTransferBudget)
-                .withUpdatedWageBudget(updatedWageBudget)
-                .withUpdatedManagerFunds(totalFunds)
-                .build();
-
-        when(fileUploadService.doesFileExist(eq(incomingClub.getLogo()))).thenReturn(true);
-
-        // execute
-        assertThrows(ServiceException.class,
-                () -> clubResource.updateClub(userPrincipal, existingClubId, incomingClub));
-
-        // assert
-        verify(clubService).getClub(any(), any());
-        verify(clubService, never()).updateClub(any(), any(), any());
-    }
-
-    /**
      * given a valid club ID, tests that the club data is removed and a 204 No Content response is returned
      */
     @Test
